pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "officerndapilib"
version = "0.0.4"
authors = [
    { name="Gibran Dar", email="gibrandar@gmail.com" }
]
description = "A thin wrapper around the OfficeRnD API"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities"
]
dependencies = [
    "attrs==23.2.0",
    "certifi==2023.11.17",
    "charset-normalizer==3.3.2",
    "colorama==0.4.6",
    "idna==3.6",
    "iniconfig==2.0.0",
    "packaging==23.2",
    "pluggy==1.3.0",
    "pytest==7.4.4",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "urllib3==2.1.0",
]

[project.urls]
"Homepage" = "https://github.com/GibranDar/officernd-api-lib"
"Bug Tracker" = "https://github.com/GibranDar/officernd-api-lib/issues"

